split Arm into 2 classes. switched to default PID
